pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personal-rss-updater"
version = "0.1.0"
description = "A robust Python program that monitors blogs for new posts and sends daily email digest notifications"
authors = [
    {name = "Yonatan Lourie", email = "yonatanlou@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pydantic[email]>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
rss-updater = "rss_updater.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rss_updater"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=rss_updater --cov-report=term-missing"
